new changes

+++ bigquery_synching/LiveMismatch/LIVE_CLC_MISMATCH.sh
@@ -1,11 +1,9 @@
 #!/bin/bash
 
-export LC_ALL=en_US.UTF-8
-export LANG=en_US.UTF-8
-export LANGUAGE=en_US.UTF-8
+set -e  # Exit immediately if any command exits with a non-zero status
+TIMEOUT_PERIOD=40
 
-
-rm -rf final_mismatch.json
+rm -f final_mismatch.json
 
 gcloud config set project prioticket-reporting
 # bq query --use_legacy_sql=False --format=prettyjson \
@@ -17,15 +15,6 @@ offset=0
 i=1
 noOfdays=$2
 
-# Check if the required arguments are provided
-if [ -z "$1" ] || [ -z "$2" ]; then
-    echo "Error: Both arguments are required."
-    echo "Usage: $0 <arg1> <arg2>"
-    exit 1
-fi
-
-echo "condition Satisfy with Both Argument"
-
 SELECTCOLUMNS="channel_level_commission_id, ticket_id, ticketpriceschedule_id, last_modified_at"
 # previous_date=$(date -d "yesterday" +"%Y-%m-%d 00:00:01")
 previous_date=$(date -d "$noOfdays days ago" +"%Y-%m-%d 00:00:01")
@@ -36,19 +25,41 @@ echo "$previous_date" >> rattan.txt
 # DBUSER="pip"
 # DBPWD="pip2024##"
 # DBDATABASE="priopassdb"
-DBHOST='163.47.214.30'
-DBUSER='datalook'
-DBPWD='datalook2024$$'
-DBDATABASE='priopassdb'
-PORT="3307"
-# DBHOST='production-primary-db-node-cluster.cluster-ck6w2al7sgpk.eu-west-1.rds.amazonaws.com'
-# DBUSER='pipeuser'
-# DBPWD='d4fb46eccNRAL'
-# DBDATABASE='priopassdb'
-# PORT="3306"
 UploadData=$1
+DATABSETYPE=$3
+
+# Check if the required arguments are provided
+if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ]; then
+    echo "Error: All arguments are required."
+    echo "Usage: $0 UploadData NoOfDays datasetType"
+    exit 1
+fi
+
+echo "condition Satisfy with Both Argument"
 
 rm -f final_mismatch.json
+
+if [[ $DATABSETYPE == "PROD" ]]; then
+
+    echo "Live Database Selected"
+
+    DBHOST='production-primary-db-node-cluster.cluster-ck6w2al7sgpk.eu-west-1.rds.amazonaws.com'
+    DBUSER='pipeuser'
+    DBPWD='d4fb46eccNRAL'
+    DBDATABASE='priopassdb'
+    PORT="3306"
+
+else
+
+    echo "Staging Database Selected"
+
+    DBHOST='163.47.214.30'
+    DBUSER='datalook'
+    DBPWD='datalook2024$$'
+    DBDATABASE='priopassdb'
+    PORT="3307"
+
+fi
 
 if [[ $UploadData == 2 ]]; then
 
@@ -58,7 +69,7 @@ if [[ $UploadData == 2 ]]; then
 
     do
 
-    RecordCount=$(mysql -h $DBHOST --user=$DBUSER --port=$PORT --password=$DBPWD $DBDATABASE -N -e "select count(*) from (select * from channel_level_commission where deleted = '0' and last_modified_at > '$previous_date' limit $offset, $limit) as base;")
+    RecordCount=$(timeout $TIMEOUT_PERIOD time mysql -h $DBHOST --user=$DBUSER --port=$PORT --password=$DBPWD $DBDATABASE -N -e "select count(*) from (select * from channel_level_commission where deleted = '0' and last_modified_at > '$previous_date' limit $offset, $limit) as base;") || exit 1
 
     echo $RecordCount
             
@@ -74,7 +85,7 @@ if [[ $UploadData == 2 ]]; then
     echo "Record from database Found offset : $offset and limit : $limit"
 
 
-    echo "select $SELECTCOLUMNS from channel_level_commission where deleted = '0' and last_modified_at > '$previous_date' limit $offset, $limit" | time mysqlsh --sql --json --uri $DBUSER@$DBHOST:$PORT -p$DBPWD --database=$DBDATABASE >> "$offset"_primarypt.json
+    echo "select $SELECTCOLUMNS from channel_level_commission where deleted = '0' and last_modified_at > '$previous_date' limit $offset, $limit" | timeout $TIMEOUT_PERIOD time mysqlsh --sql --json --uri $DBUSER@$DBHOST:$PORT -p$DBPWD --database=$DBDATABASE >> "$offset"_primarypt.json || exit 1
 
     jq 'select(.warning | not)' "$offset"_primarypt.json >> "$offset"_primarypt1.json
 
@@ -99,12 +110,12 @@ else
 
 
     bq query --use_legacy_sql=False --max_rows=1000000 --format=prettyjson \
-    "with tlc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_test.channel_level_mismatch), tlc as (select * from tlc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), clc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_olap.channel_level_commission), clc as (select * from clc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), base as (SELECT tlc.*, clc.channel_level_commission_id as id FROM tlc left join  clc on tlc.channel_level_commission_id = clc.channel_level_commission_id and tlc.ticket_id = clc.ticket_id and tlc.ticketpriceschedule_id = clc.ticketpriceschedule_id and (tlc.last_modified_at = clc.last_modified_at or tlc.last_modified_at <= clc.last_modified_at)) select * from base where id is NULL" > final_mismatch.json
+    "with tlc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_test.channel_level_mismatch), tlc as (select * from tlc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), clc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_olap.channel_level_commission), clc as (select * from clc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), base as (SELECT tlc.*, clc.channel_level_commission_id as id FROM tlc left join  clc on tlc.channel_level_commission_id = clc.channel_level_commission_id and tlc.ticket_id = clc.ticket_id and tlc.ticketpriceschedule_id = clc.ticketpriceschedule_id and (tlc.last_modified_at = clc.last_modified_at or tlc.last_modified_at <= clc.last_modified_at)) select * from base where id is NULL" > final_mismatch.json || exit 1
 
 fi
 
 
 bq query --use_legacy_sql=False --max_rows=1000000 --format=prettyjson \
-"with tlc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_test.channel_level_mismatch), tlc as (select * from tlc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), clc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_olap.channel_level_commission), clc as (select * from clc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), base as (SELECT tlc.*, clc.channel_level_commission_id as id FROM tlc left join  clc on tlc.channel_level_commission_id = clc.channel_level_commission_id and tlc.ticket_id = clc.ticket_id and tlc.ticketpriceschedule_id = clc.ticketpriceschedule_id and (tlc.last_modified_at = clc.last_modified_at or tlc.last_modified_at <= clc.last_modified_at)) select * from base where id is NULL" > final_mismatch.json
+"with tlc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_test.channel_level_mismatch), tlc as (select * from tlc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), clc1 as (select *,row_number() over(partition by channel_level_commission_id order by last_modified_at desc ) as rn from prioticket-reporting.prio_olap.channel_level_commission), clc as (select * from clc1 where rn=1 and last_modified_at > TIMESTAMP(CONCAT(CAST(DATE(TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL $noOfdays DAY)) AS STRING), ' 00:00:00')) AND last_modified_at <= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 45 MINUTE)), base as (SELECT tlc.*, clc.channel_level_commission_id as id FROM tlc left join  clc on tlc.channel_level_commission_id = clc.channel_level_commission_id and tlc.ticket_id = clc.ticket_id and tlc.ticketpriceschedule_id = clc.ticketpriceschedule_id and (tlc.last_modified_at = clc.last_modified_at or tlc.last_modified_at <= clc.last_modified_at)) select * from base where id is NULL" > final_mismatch.json || exit 1
 
 source update_commission_CLC.sh
